Created UserInput Class
I created a UserInput Class that includes two methods: readNextLine and readNextInt. The first method is for reading strings, and the second method is for reading integers. This class will be implemented wherever a Scanner is needed. It helps centralize user input and validate it with try-catch blocks and exceptions. Also allows user to exit your program at any time by typing the word, "exit" (case insensitive).