Refactor input handling in Student class with UserInput methods
I replaced Scanner input with UserInput methods for better input handling. This allowed me to merge a lot of the print statements with the input handling. It also allowed me to remove the Scanner import. I decided to only refactor input handling in the Student class. I am aware that changes will have to be made to the other class (ex. FullTimeStudent) as they also deal with scanner objects. But I only changed the Student class as an example since there is a lot of code that needs to be changed.